[HOTFIX] 배포 문제 해결
- formatDate 수정
- 로그인 안됐을 때 login으로 라우팅
